pyproject.toml
[project]
name = "bastuflotten"
version = "0.1.0"
description = "Slack booking bot for Bastuflotten"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "slack-bolt>=1.21.0",
    "slack-sdk>=3.31.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.13.5",
    "certifi>=2026.2.25",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bastuflotten"]
